Refactor datagrid
