+++ packages/perspective-viewer-datagrid/src/js/model/index.js
@@ -1,0 +1,10 @@
+/******************************************************************************
+ *
+ * Copyright (c) 2017, the Perspective Authors.
+ *
+ * This file is part of the Perspective library, distributed under the terms of
+ * the Apache License 2.0.  The full license can be found in the LICENSE file.
+ *
+ */
+
+export const PRIVATE_PLUGIN_SYMBOL = Symbol("Plugin Symbol");
