--- packages/perspective-viewer-datagrid/src/js/plugin.js
@@ -1,433 +1,0 @@
-/******************************************************************************
- *
- * Copyright (c) 2017, the Perspective Authors.
- *
- * This file is part of the Perspective library, distributed under the terms of
- * the Apache License 2.0.  The full license can be found in the LICENSE file.
- *
- */
-
-import "regular-table";
-
-import {
-    createModel,
-    configureRegularTable,
-    formatters,
-} from "./regular_table_handlers.js";
-
-import MATERIAL_STYLE from "../less/regular_table.less";
-import TOOLBAR_STYLE from "../less/toolbar.less";
-import {configureRowSelectable, deselect} from "./row_selection.js";
-import {configureClick} from "./click.js";
-import {configureEditable} from "./editing.js";
-import {configureSortable} from "./sorting.js";
-import {PLUGIN_SYMBOL, make_color_record} from "./plugin_menu.js";
-
-export class HTMLPerspectiveViewerDatagridPluginElement extends HTMLElement {
-    constructor() {
-        super();
-        this.regular_table = document.createElement("regular-table");
-        this.regular_table.formatters = formatters;
-        this._is_scroll_lock = true;
-    }
-
-    connectedCallback() {
-        if (!this._toolbar) {
-            this._toolbar = document.createElement(
-                "perspective-viewer-datagrid-toolbar"
-            );
-
-            this._toolbar.setAttribute("slot", "plugin-settings");
-            this._toolbar.attachShadow({mode: "open"});
-            this._toolbar.shadowRoot.innerHTML = `
-                <style>
-                    ${TOOLBAR_STYLE}
-                </style>
-                <div id="toolbar">
-                    <span id="scroll_lock" class="lock-scroll button">
-                        <span>Align Scroll</span>
-                    </span>
-                    <span id="edit_mode" class="button"><span>Read Only</span></span>
-                </div>
-            `;
-
-            this._scroll_lock =
-                this._toolbar.shadowRoot.querySelector("#scroll_lock");
-
-            this._scroll_lock.addEventListener("click", () =>
-                this._toggle_scroll_lock()
-            );
-
-            this._edit_mode =
-                this._toolbar.shadowRoot.querySelector("#edit_mode");
-
-            this._edit_mode.addEventListener("click", () => {
-                this._toggle_edit_mode();
-                this.regular_table.draw();
-                this.parentElement.dispatchEvent(
-                    new Event("perspective-config-update")
-                );
-            });
-        }
-
-        this.parentElement.appendChild(this._toolbar);
-    }
-
-    disconnectedCallback() {
-        this._toolbar.parentElement.removeChild(this._toolbar);
-    }
-
-    _toggle_edit_mode(force = undefined) {
-        if (typeof force === "undefined") {
-            force = !this._is_edit_mode;
-        }
-
-        this._is_edit_mode = force;
-        this.classList.toggle("editable", force);
-        if (this._edit_mode !== undefined) {
-            this._edit_mode.classList.toggle("editable", force);
-            if (force) {
-                this._edit_mode.children[0].textContent = "Editable";
-            } else {
-                this._edit_mode.children[0].textContent = "Read Only";
-            }
-        }
-    }
-
-    _toggle_scroll_lock(force = undefined) {
-        if (typeof force === "undefined") {
-            force = !this._is_scroll_lock;
-        }
-
-        this._is_scroll_lock = force;
-        this.classList.toggle("sub-cell-scroll-enabled", !force);
-        if (this._scroll_lock !== undefined) {
-            this._scroll_lock.classList.toggle("lock-scroll", force);
-            if (!force) {
-                this._scroll_lock.children[0].textContent = "Free Scroll";
-            } else {
-                this._scroll_lock.children[0].textContent = "Align Scroll";
-            }
-        }
-    }
-
-    async activate(view) {
-        let viewer = this.parentElement;
-        let table = await viewer.getTable(true);
-        if (!this._initialized) {
-            this.innerHTML = "";
-            this.appendChild(this.regular_table);
-            this.model = await createModel(this.regular_table, table, view);
-            configureRegularTable(this.regular_table, this.model);
-            await configureRowSelectable.call(
-                this.model,
-                this.regular_table,
-                viewer
-            );
-            await configureClick.call(this.model, this.regular_table, viewer);
-            await configureEditable.call(
-                this.model,
-                this.regular_table,
-                viewer
-            );
-            await configureSortable.call(
-                this.model,
-                this.regular_table,
-                viewer
-            );
-            this._initialized = true;
-        } else {
-            await createModel(this.regular_table, table, view, this.model);
-        }
-    }
-
-    get name() {
-        return "Datagrid";
-    }
-
-    get category() {
-        return "Basic";
-    }
-
-    get select_mode() {
-        return "toggle";
-    }
-
-    get min_config_columns() {
-        return undefined;
-    }
-
-    get config_column_names() {
-        return undefined;
-    }
-
-    async draw(view) {
-        if (this.parentElement) {
-            await this.activate(view);
-        }
-
-        if (!this.isConnected || this.offsetParent == null) {
-            return;
-        }
-
-        const old_sizes = this._save_column_size_overrides();
-        let viewer = this.parentElement;
-        const draw = this.regular_table.draw({invalid_columns: true});
-        if (!this.model._preserve_focus_state) {
-            this.regular_table.scrollTop = 0;
-            this.regular_table.scrollLeft = 0;
-            deselect(this.regular_table, viewer);
-            this.regular_table._resetAutoSize();
-        } else {
-            this.model._preserve_focus_state = false;
-        }
-
-        this._restore_column_size_overrides(old_sizes);
-        await draw;
-
-        this._toolbar.classList.toggle(
-            "aggregated",
-            this.model._config.group_by.length > 0 ||
-                this.model._config.split_by.length > 0
-        );
-    }
-
-    async update(view) {
-        this.model._num_rows = await view.num_rows();
-        await this.regular_table.draw();
-    }
-
-    async resize() {
-        if (!this.isConnected || this.offsetParent == null) {
-            return;
-        }
-
-        if (this._initialized) {
-            await this.regular_table.draw();
-        }
-    }
-
-    async clear() {
-        this.regular_table._resetAutoSize();
-        this.regular_table.clear();
-    }
-
-    save() {
-        if (this.regular_table) {
-            const datagrid = this.regular_table;
-            const token = {
-                columns: {},
-                scroll_lock: !!this._is_scroll_lock,
-                editable: !!this._is_edit_mode,
-            };
-
-            for (const col of Object.keys(datagrid[PLUGIN_SYMBOL] || {})) {
-                const config = Object.assign({}, datagrid[PLUGIN_SYMBOL][col]);
-                if (config?.pos_fg_color || config?.pos_bg_color) {
-                    config.pos_fg_color = config.pos_fg_color?.[0];
-                    config.neg_fg_color = config.neg_fg_color?.[0];
-                    config.pos_bg_color = config.pos_bg_color?.[0];
-                    config.neg_bg_color = config.neg_bg_color?.[0];
-                }
-
-                if (config?.color) {
-                    config.color = config.color[0];
-                }
-
-                token.columns[col] = config;
-            }
-
-            const column_size_overrides = this._save_column_size_overrides();
-
-            for (const col of Object.keys(column_size_overrides || {})) {
-                if (!token.columns[col]) {
-                    token.columns[col] = {};
-                }
-
-                token.columns[col].column_size_override =
-                    column_size_overrides[col];
-            }
-
-            return JSON.parse(JSON.stringify(token));
-        }
-        return {};
-    }
-
-    restore(token) {
-        token = JSON.parse(JSON.stringify(token));
-        const overrides = {};
-        if (token.columns) {
-            for (const col of Object.keys(token.columns)) {
-                const col_config = token.columns[col];
-                if (col_config.column_size_override !== undefined) {
-                    overrides[col] = col_config.column_size_override;
-                    delete col_config["column_size_override"];
-                }
-
-                if (col_config?.pos_fg_color) {
-                    col_config.pos_fg_color = make_color_record(
-                        col_config.pos_fg_color
-                    );
-                    col_config.neg_fg_color = make_color_record(
-                        col_config.neg_fg_color
-                    );
-                }
-
-                if (col_config?.pos_bg_color) {
-                    col_config.pos_bg_color = make_color_record(
-                        col_config.pos_bg_color
-                    );
-                    col_config.neg_bg_color = make_color_record(
-                        col_config.neg_bg_color
-                    );
-                }
-
-                if (col_config?.color) {
-                    col_config.color = make_color_record(col_config.color);
-                }
-
-                if (Object.keys(col_config).length === 0) {
-                    delete token.columns[col];
-                }
-            }
-        }
-
-        if ("editable" in token) {
-            this._toggle_edit_mode(token.editable);
-        }
-
-        if ("scroll_lock" in token) {
-            this._toggle_scroll_lock(token.scroll_lock);
-        }
-
-        const datagrid = this.regular_table;
-        try {
-            datagrid._resetAutoSize();
-        } catch (e) {
-            // Do nothing;  this may fail if no auto size info has been read.
-            // TODO fix this regular-table API
-        }
-
-        this._restore_column_size_overrides(overrides, true);
-        datagrid[PLUGIN_SYMBOL] = token.columns;
-    }
-
-    async restyle(view) {
-        this.draw(view);
-    }
-
-    delete() {
-        if (this.regular_table.table_model) {
-            this.regular_table._resetAutoSize();
-        }
-        this.regular_table.clear();
-    }
-
-    // Private
-
-    /**
-     * Extract the current user-overriden column widths from
-     * `regular-table`.  This functiond depends on the internal
-     * implementation of `regular-table` and may break!
-     *
-     * @returns An Object-as-dictionary keyed by column_path string, and
-     * valued by the column's user-overridden pixel width.
-     */
-    _save_column_size_overrides() {
-        if (!this._initialized) {
-            return [];
-        }
-
-        if (this._cached_column_sizes) {
-            const x = this._cached_column_sizes;
-            this._cached_column_sizes = undefined;
-            return x;
-        }
-
-        const overrides = this.regular_table._column_sizes.override;
-        const {group_by, columns} = this.model._config;
-        const tree_header_offset =
-            group_by?.length > 0 ? group_by.length + 1 : 0;
-
-        const old_sizes = {};
-        for (const key of Object.keys(overrides)) {
-            if (overrides[key] !== undefined) {
-                const index = key - tree_header_offset;
-                if (index > -1) {
-                    old_sizes[this.model._column_paths[index]] = overrides[key];
-                } else if (index === -1) {
-                    old_sizes["__ROW_PATH__"] = overrides[key];
-                }
-            }
-        }
-
-        return old_sizes;
-    }
-
-    /**
-     * Restore a saved column width override token.
-     *
-     * @param {*} token An object previously returned by a call to
-     * `_save_column_size_overrides()`
-     * @param {*} [cache=false] A flag indicating whether this value should
-     * be cached so a future `resetAutoSize()` call does not clear it.
-     * @returns
-     */
-    _restore_column_size_overrides(old_sizes, cache = false) {
-        if (!this._initialized) {
-            return;
-        }
-
-        if (cache) {
-            this._cached_column_sizes = old_sizes;
-        }
-
-        const overrides = {};
-        const {group_by, columns} = this.model._config;
-        const tree_header_offset =
-            group_by?.length > 0 ? group_by.length + 1 : 0;
-
-        for (const key of Object.keys(old_sizes)) {
-            if (key === "__ROW_PATH__") {
-                overrides[tree_header_offset - 1] = old_sizes[key];
-            } else {
-                const index = this.model._column_paths.indexOf(key);
-                overrides[index + tree_header_offset] = old_sizes[key];
-            }
-        }
-
-        this.regular_table._column_sizes.override = overrides;
-    }
-}
-
-customElements.define(
-    "perspective-viewer-datagrid",
-    HTMLPerspectiveViewerDatagridPluginElement
-);
-
-/**
- * Appends the default table CSS to `<head>`, should be run once on module
- * import.
- *
- */
-function _register_global_styles() {
-    const style = document.createElement("style");
-    style.textContent = MATERIAL_STYLE;
-    document.head.insertBefore(style, document.head.firstChild);
-}
-
-/******************************************************************************
- *
- * Main
- *
- */
-
-async function _register_element() {
-    await customElements.whenDefined("perspective-viewer");
-    customElements
-        .get("perspective-viewer")
-        .registerPlugin("perspective-viewer-datagrid");
-}
-
-_register_element();
-_register_global_styles();
