+++ packages/perspective-viewer-datagrid/src/js/style_handlers/table_cell/cell_flash.js
