+++ packages/perspective-viewer-datagrid/src/js/style_handlers/table_cell/datetime.js
@@ -1,0 +1,13 @@
+/******************************************************************************
+ *
+ * Copyright (c) 2017, the Perspective Authors.
+ *
+ * This file is part of the Perspective library, distributed under the terms of
+ * the Apache License 2.0.  The full license can be found in the LICENSE file.
+ *
+ */
+
+export function cell_style_datetime(plugin, td, metadata) {
+    td.style.backgroundColor = "";
+    td.style.color = "";
+}
