+++ packages/perspective-viewer-datagrid/test/results/results.json
@@ -11,21 +11,21 @@
     "superstore_filters_by_a_datetime_column_": "4f3ebe78d2092c8c6f74389778d764d5",
     "superstore_highlights_invalid_filter_": "d74bde78c8afbd835e451fd5afffc9fe",
     "superstore_sorts_by_an_alpha_column_": "e404bb87320c31ec31a54c8652ae439c",
-    "superstore_displays_visible_columns_": "0b95713a81bc8af9b2db97111cdba9fa",
+    "superstore_displays_visible_columns_": "30dfb2e0551c4f598166959f5d7b3673",
     "superstore_resets_viewable_area_when_the_logical_size_expands_": "e9dd1d275f46f6a0857e0168835d0b38",
     "superstore_resets_viewable_area_when_the_physical_size_expands_": "e469597235cc032619bc095a76f9ad6f",
-    "__GIT_COMMIT__": "051dbb89de8bae97970b5988950cfcb41a1e51c8",
-    "superstore_shows_a_grid_without_any_settings_applied": "2bf7d29b6483e16ad1ff65a1814f2230",
-    "superstore_pivot_by_a_row": "293fedf936989ef0a3f05d2e06b9ab97",
-    "superstore_pivot_by_two_rows": "ed33f1f97a87fffecf9de84db6130af6",
+    "__GIT_COMMIT__": "0ee79b07e87340cddb5626dfc732ca85932f6d7c",
+    "superstore_shows_a_grid_without_any_settings_applied": "8ec13290bdbbe3aa17405fc37fe771b7",
+    "superstore_pivot_by_a_row": "34f5a466415985ee75aeebcf3dc4f914",
+    "superstore_pivot_by_two_rows": "02dd963e15fa30e2a49a37fe3b606638",
     "superstore_pivot_by_a_column": "13d1dbd9b8fa97a9053b111a93444d87",
-    "superstore_pivot_by_a_row_and_a_column": "471ee7620231192592a17ac737197295",
-    "superstore_pivot_by_two_rows_and_two_columns": "7bdeb8367e323bd7d9927ee069a45851",
-    "superstore_sort_by_a_hidden_column": "a88f622d32cd13a6b1c330668eabd9ac",
-    "superstore_sort_by_a_numeric_column": "db67040f6d638a0c889c0cffb3832057",
-    "superstore_sort_by_an_alpha_column": "9e7a551308497ddca3e1b058cc69ba8c",
-    "superstore_filters_filters_by_a_numeric_column": "ce3bed9620c8fde9929d3a3ed548279c",
-    "superstore_filters_filters_by_an_alpha_column": "5cac36a354a18935179afe6b8c3c10da",
-    "superstore_filters_filters_with__in__comparator": "d7047b2da9afe8df7e7f8ceebea13f43",
-    "superstore_perspective-config-update_event_is_fired_when_column_style_is_changed": "8aa6557c21232f592904e73cf26c6675"
+    "superstore_pivot_by_a_row_and_a_column": "f51e9465b6873f1b3e92aacb43ec89b8",
+    "superstore_pivot_by_two_rows_and_two_columns": "b5a05fc5031782f5b4d5d3f13dc2c966",
+    "superstore_sort_by_a_hidden_column": "80cfa97284910fa193e34a18286be57f",
+    "superstore_sort_by_a_numeric_column": "fa82c95d2f24795a11dfdaf7699511fb",
+    "superstore_sort_by_an_alpha_column": "e9f3d537de1c6283391623749974a2e8",
+    "superstore_filters_filters_by_a_numeric_column": "4e4f33833fb531bbdff6fe61694befd8",
+    "superstore_filters_filters_by_an_alpha_column": "2ee0b0c9c37540db277cfe6d390ebf2b",
+    "superstore_filters_filters_with__in__comparator": "414732a1eedbc4a3fcf070451e2aa7ed",
+    "superstore_perspective-config-update_event_is_fired_when_column_style_is_changed": "7885342bae75f6796eb231db2b741eeb"
 }
