+++ packages/perspective-workspace/test/results/results.json
@@ -1,17 +1,17 @@
 {
-    "__GIT_COMMIT__": "36b408fd2f211dc36d62ce6fc233f061ede18c69",
+    "__GIT_COMMIT__": "c855e38b34e3ec514c2e8b61d1807ee89231e348",
     "index_restore_workspace_with_detail_only": "d24f601369fbf86c853d4dd2894506e3",
-    "index_Light_DOM_restore_workspace_with_detail_only": "82be3479bd7fb25462244a1a93d1b7dd",
+    "index_Light_DOM_restore_workspace_with_detail_only": "a86ed09f48f11f42776af77944fda2cf",
     "index_Shadow_DOM_restore_workspace_with_detail_only": "087c2dffe30a52fa4ea1b127ce83c625",
-    "index_Light_DOM_restore_workspace_with_master_and_detail": "7818f24ea922c7c0981c3dd4b69943a1",
+    "index_Light_DOM_restore_workspace_with_master_and_detail": "3fbd38e883ea2310419508607b81a69d",
     "index_Shadow_DOM_restore_workspace_with_master_and_detail": "34a8933df88660871ef3ca3e1e2d6a11",
     "index_Light_DOM_restore_workspace_with_viewers_with_generated_slotids": "0fad47e40121ecccf41c21bdfe2b7d56",
     "index_Shadow_DOM_restore_workspace_with_viewers_with_generated_slotids": "606bea2473f0e8b91756d5b69cd982be",
-    "index_Light_DOM_removeChild_Remove_One": "b29f031ce06cba2bc775ed1775683170",
+    "index_Light_DOM_removeChild_Remove_One": "39ee7d3dd2cbff6f4ff9d960633b6fb2",
     "index_Shadow_DOM_removeChild_Remove_One": "8f1a715acfc9a0415b9400a070dedfa4",
-    "index_Light_DOM_Create_One": "52d741eaacaf1c7f6a53e56ec40a5978",
-    "index_Light_DOM_Create_Multiple": "b5a5fc86319ea59cb80168351426d331",
-    "index_Light_DOM_Create_multiple_with_names": "18c4c3840e7f131281b6062cd0d3b4af",
+    "index_Light_DOM_Create_One": "49010bc78a8a2375e57a2aa1627471ab",
+    "index_Light_DOM_Create_Multiple": "c8b96c9f843a022aa4f8b77a0a28312c",
+    "index_Light_DOM_Create_multiple_with_names": "a2dcd3235172a4921858273a52b5855a",
     "index_Shadow_DOM_Create_One": "9731261b7a34c19fcb11855daf723c73",
     "index_Shadow_DOM_Create_Multiple": "2776b343d54425badd479d57d300233e",
     "index_Shadow_DOM_Create_multiple_with_names": "e8bd434eb1637078da674adf08e1d8dc",
